requirements.txt
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
pandas==2.2.3
pillow==10.3.0
scikit-image==0.24.0
scikit-learn==1.4.2
seaborn==0.13.2
tqdm==4.67.1